ensure supervisord and plex library paths exist before trying to execute supervisord.  if plex uid/gid defined use that for library

+++ assets/scripts/startup.sh
@@ -2,6 +2,10 @@
 
 # directory to store plex build downloads
 DOWNLOAD_DIR="/plex/downloads"
+# plex library
+PLEX_LIBRARY="/plex/Library"
+# supervisord logs
+SUPERVISORD_LOGS="/plex/logs/supervisor"
 
 # number of previous releases to keep in download directory
 PLEX_NUM=2
@@ -21,6 +25,7 @@ fi
 # if not, download it
 if [ ! -f "${DOWNLOAD_DIR}/plexmediaserver_${PLEX_SERVER_VERSION}.deb" ]; then
   # download plex media server
+  echo "downloading plexmediaserver_${PLEX_SERVER_VERSION}"
   wget -O ${DOWNLOAD_DIR}/plexmediaserver_${PLEX_SERVER_VERSION}.deb \
        -q https://downloads.plex.tv/plex-media-server/${PLEX_SERVER_VERSION}/plexmediaserver_${PLEX_SERVER_VERSION}_${PLEX_SERVER_ARCH}.deb
 fi
@@ -40,7 +45,35 @@ for BUILD in $(find ${DOWNLOAD_DIR} -name plexmediaserver*.deb -print0| xargs -0
 done
 
 # update default config file
-mv -f /tmp/default-plexmediaserver /etc/default/plexmediaserver
+if [ -f /tmp/default-plexmediaserver ]; then
+  echo "copying config to /etc/default/plexmediaserver"
+  mv -f /tmp/default-plexmediaserver /etc/default/plexmediaserver
+fi
+
+# force update plex user to match NFS user if defined
+if [[ (! -z ${PLEX_UID}) && (! -z ${PLEX_GID}) ]]; then
+  echo "setting up PLEX UID:${PLEX_UID} and GID:${PLEX_GID}"
+  groupmod -g ${PLEX_GID} plex
+  usermod -u ${PLEX_UID} plex
+fi
+
+# preseed plex library and ensure permissions are setup properly
+if [ ! -d ${PLEX_LIBRARY} ]; then
+  echo "setting up PLEX Library"
+  mkdir -p -m 2775 ${PLEX_LIBRARY}
+  chown -R plex:plex ${PLEX_LIBRARY}
+else
+  # ensure permissions are correct if we exist
+  chown -R plex:plex ${PLEX_LIBRARY}
+  chmod 2775 ${PLEX_LIBRARY}
+fi
+
+# ensure supervisor logfile is present
+if [ ! -f ${SUPERVISORD_LOGS}/plex.log ]; then
+  echo "setting up ${SUPERVISORD_LOGS}"
+  mkdir -p ${SUPERVISORD_LOGS}
+  touch ${SUPERVISORD_LOGS}/plex.log
+fi
 
 # start supervisor
 /usr/bin/supervisord -n -c /etc/supervisor/supervisord.conf
